requirements.txt
langchain
openai
pinecone-client
python-dotenv==1.0.0
uuid
pathlib
uvicorn
fastapi
jinja2
slowapi
httpx
git+https://github.com/casics/nostril.git
tiktoken
langsmith
bs4
apscheduler
torch
transformers
sentencepiece
protobuf
cohere